UNOMI-263 - removed commands in favour of datafetchers
